+++ quantify_scheduler/backends/types/qblox.py
@@ -443,8 +443,10 @@ class SequencerSettings(DataClassJsonMixin):
                 )
             return val
 
-        modulation_freq: Union[float, None] = seq_settings.get("interm_freq", None)
-        nco_en: bool = modulation_freq != 0 and modulation_freq is not None
+        modulation_freq: Optional[float] = seq_settings.get("interm_freq", None)
+        nco_en: bool = (
+            modulation_freq is not None and modulation_freq != 0
+        )  # Allow NCO to be permanently disabled via `"interm_freq": 0` in the hardware config
 
         mixer_amp_ratio = extract_and_verify_range(
             param_name="mixer_amp_ratio",